backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.0
orjson>=3.9.15